Implemented search bar and pagination.

+++ frontend/src/components/CreatorCard.js
@@ -9,24 +9,22 @@ function CreatorCard({ creator, onDelete }) {
   const navigate = useNavigate();
 
   const handleDelete = async () => {
-    const { error } = await supabase.from('creators').delete().eq('id', creator.id);
-    if (error) {
-      console.error('Error deleting creator:', error);
-    } else {
-      onDelete(creator.id);
+    const confirmed = window.confirm("Are you sure you want to delete this creator?");
+    if (confirmed) {
+      const { error } = await supabase.from('creators').delete().eq('id', creator.id);
+      if (error) {
+        console.error('Error deleting creator:', error);
+      } else {
+        onDelete(creator.id);
+      }
     }
   };
 
-  // Function to truncate the description
   const truncateDescription = (description, maxLength) => {
-    if (description.length <= maxLength) {
-      return description;
-    }
-    return description.slice(0, maxLength) + '...';
+    return description.length <= maxLength ? description : description.slice(0, maxLength) + '...';
   };
 
   const handleCardClick = (e) => {
-    // Prevent navigation if an action button is clicked
     if (e.target.closest('button') || e.target.closest('a')) return;
     navigate(`/creator/${creator.id}`);
   };
@@ -37,17 +35,18 @@ function CreatorCard({ creator, onDelete }) {
       <header>
         <h3 className="card-text">{creator.name}</h3>
       </header>
-      <p className="card-text">{truncateDescription(creator.description, 100)}</p> {/* Truncate description to 100 characters */}
+      <p className="card-text">{truncateDescription(creator.description, 100)}</p>
       <footer>
         <a href={creator.url} target="_blank" rel="noopener noreferrer" className="icon-button" onClick={(e) => e.stopPropagation()}>
           <i className="fas fa-link"></i>
+          <span className="icon-text">Channel</span>
         </a>
-        
         <Link to={`/edit/${creator.id}`} className="icon-button" onClick={(e) => e.stopPropagation()}>
           <i className="fas fa-pen"></i>
+          <span className="icon-text">Edit</span>
         </Link>
         <button onClick={(e) => { e.stopPropagation(); handleDelete(); }} className="icon-button delete-button">
-          <i className="fas fa-times"></i>
+          <i className="fas fa-times delete-icon"></i>
         </button>
       </footer>
     </article>
